+++ tests/testthat/test-setup_trial.R
@@ -111,6 +111,7 @@ test_that("validate setup trial specifications", {
     fun_draws = adaptr:::get_draws_binom,
     fun_raw_est = mean,
     min_probs = rep(0.15, 3),
+    rescale_probs = "limits",
     data_looks = seq(from = 300, to = 2000, by = 100),
     equivalence_prob = 0.9,
     equivalence_diff = 0.05,
@@ -129,6 +130,7 @@ test_that("validate setup trial specifications", {
     control = "B",
     true_ys = c(0.25, 0.20, 0.30),
     min_probs = rep(0.15, 3),
+    rescale_probs = "limits",
     data_looks = seq(from = 300, to = 2000, by = 100),
     equivalence_prob = 0.9,
     equivalence_diff = 0.05,
@@ -231,6 +233,15 @@ test_that("setup/validate_trial functions errors on invalid inputs", {
                               futility_prob = 0.9, futility_diff = 0.1 * 1:3, futility_only_first = TRUE))
   expect_error(validate_trial(arms = 1:3, control = 1, data_looks = 1:3 * 100,
                               futility_prob = 0.9, futility_diff = 0.1, futility_only_first = NA))
+
+  expect_error(validate_trial(arms = 1:2, data_looks = 1:3 * 100,
+                              rescale_adapt_probs = "both"))
+  expect_error(validate_trial(arms = 1:3, control = 1, data_looks = 1:3 * 100,
+                              rescale_adapt_probs = "both"))
+  expect_error(validate_trial(arms = 1:3, data_looks = 1:3 * 100,
+                              rescale_adapt_probs = "invalid"))
+  expect_error(validate_trial(arms = 1:3, data_looks = 1:3 * 100,
+                              rescale_adapt_probs = c("superiority", "inferiority")))
 
   expect_error(validate_trial(arms = 1:3, data_looks = 1:3 * 100, soften_power = 1 - 0.01 * 1:2))
   expect_error(validate_trial(arms = 1:3, data_looks = 1:3 * 100, soften_power = 1.01))