+++ tests/testthat/test-update_saved.R
@@ -11,6 +11,7 @@ test_that("updating outdated trial_results objects works", {
   pseudo_old_res <- res
   pseudo_old_res$adaptr_version <- NULL # mimic what happened in adaptr until v1.1.1
   pseudo_old_res$trial_spec$rescale_probs <- NULL
+  pseudo_old_res$trial_spec$rescale_adapt_probs <- NULL
 
   saveRDS(res, tmp_file)
   expect_warning(update_saved_trials(tmp_file))
@@ -27,10 +28,19 @@ test_that("updating outdated trial_results objects works", {
   saveRDS(1:10, tmp_file)
   expect_error(update_saved_trials(tmp_file))
 
-  # Test v1.2.0+
+  # Test v1.2.0 to 1.3.2
   pseudo_old_res <- res
   pseudo_old_res$trial_spec$rescale_probs <- NULL
+  pseudo_old_res$trial_spec$rescale_adapt_probs <- NULL
   pseudo_old_res$adaptr_version <- as.package_version("1.2.0")
+
+  saveRDS(pseudo_old_res, tmp_file)
+  expect_identical(update_saved_trials(tmp_file), res)
+
+  # Test v1.4.0
+  pseudo_old_res <- res
+  pseudo_old_res$trial_spec$rescale_adapt_probs <- NULL
+  pseudo_old_res$adaptr_version <- as.package_version("1.4.0")
 
   saveRDS(pseudo_old_res, tmp_file)
   expect_identical(update_saved_trials(tmp_file), res)
@@ -45,11 +55,13 @@ test_that("updating outdated trial_calibration objects works", {
 
   res <- read_testdata("binom___calibration___setup2_arms__no_difference___rar")
 
-  # Test v1.3.0+
+  # Test v1.3.0-1.3.2 and general behaviour
   pseudo_old_res <- res
   pseudo_old_res$adaptr_version <- as.package_version("1.3.0")
   pseudo_old_res$input_trial_spec$rescale_probs <- NULL
   pseudo_old_res$best_trial_spec$rescale_probs <- NULL
+  pseudo_old_res$input_trial_spec$rescale_adapt_probs <- NULL
+  pseudo_old_res$best_trial_spec$rescale_adapt_probs <- NULL
   pseudo_old_res$best_sims$adaptr_version <- as.package_version("1.3.0")
 
   saveRDS(res, tmp_file)
@@ -58,7 +70,19 @@ test_that("updating outdated trial_calibration objects works", {
   saveRDS(pseudo_old_res, tmp_file)
   expect_invisible(update_saved_calibration(tmp_file))
   expect_warning(update_saved_calibration(tmp_file))
+  saveRDS(pseudo_old_res, tmp_file)
+  expect_identical(update_saved_calibration(tmp_file), res)
 
   saveRDS(1:10, tmp_file)
   expect_error(update_saved_calibration(tmp_file))
+
+  # Test v1.4.0 specific behaviour
+  pseudo_old_res <- res
+  pseudo_old_res$adaptr_version <- as.package_version("1.4.0")
+  pseudo_old_res$input_trial_spec$rescale_adapt_probs <- NULL
+  pseudo_old_res$best_trial_spec$rescale_adapt_probs <- NULL
+  pseudo_old_res$best_sims$adaptr_version <- as.package_version("1.4.0")
+
+  saveRDS(pseudo_old_res, tmp_file)
+  expect_identical(update_saved_calibration(tmp_file), res)
 })